Patch by Alan Young <ayoung at teleport.com>: Fix xacodec type mismatch, which is exploited on 64bit-archs (where sizeof(int) != sizeof(long))
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@6699 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ xacodec.h
@@ -29,15 +29,15 @@ typedef struct
 typedef struct
 {
   void			*anim_hdr;
-  unsigned long		compression;
-  unsigned long		x, y;
-  unsigned long		depth;
+  unsigned int		compression;
+  unsigned int		x, y;
+  unsigned int		depth;
   void			*extra;
-  unsigned long		xapi_rev;
-  unsigned long		(*decoder)();
+  unsigned int		xapi_rev;
+  unsigned int		(*decoder)();
   char			*description;
-  unsigned long		avi_ctab_flag;
-  unsigned long		(*avi_read_ext)();
+  unsigned int		avi_ctab_flag;
+  unsigned int		(*avi_read_ext)();
 } XA_CODEC_HDR;
 
 #define CODEC_SUPPORTED 1
